requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
addict=2.4.0=pypi_0
aliyun-python-sdk-core=2.16.0=pypi_0
aliyun-python-sdk-kms=2.16.5=pypi_0
blas=1.0=mkl
ca-certificates=2025.12.2=haa95532_0
certifi=2026.2.25=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.6=pypi_0
click=8.1.8=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.1.1=pypi_0
crcmod=1.7=pypi_0
cryptography=46.0.5=pypi_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.12.1=pypi_0
e2cnn=0.2.3=pypi_0
filelock=3.14.0=pypi_0
fonttools=4.57.0=pypi_0
freetype=2.14.1=hfbffc0b_0
idna=3.11=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
intel-openmp=2023.1.0=h59b6b97_46320
jmespath=0.10.0=pypi_0
jpeg=9b=hb83a4c4_2
kiwisolver=1.4.7=pypi_0
lerc=4.0.0=h5da7b33_0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libpng=1.6.55=h2854ad3_0
libtiff=4.5.0=h8a3f274_0
libuv=1.52.0=hd0d7782_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_1
libzlib=1.3.1=h02ab6af_0
lz4-c=1.9.4=h2bbff1b_1
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
matplotlib=3.7.5=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py38h2bbff1b_1
mkl_fft=1.3.8=py38h2bbff1b_0
mkl_random=1.2.4=py38h59b6b97_0
mmcv-full=1.7.2=pypi_0
mmdet=2.28.2=pypi_0
mmengine=0.10.7=pypi_0
mmrotate=0.3.4=dev_0
model-index=0.1.11=pypi_0
mpmath=1.3.0=pypi_0
ninja=1.13.1=haa95532_0
ninja-base=1.13.1=h512dff2_0
numpy=1.24.3=py38h79a8e48_1
numpy-base=1.24.3=py38h8a87ada_1
opencv-python=4.13.0.92=pypi_0
opendatalab=0.0.10=pypi_0
openmim=0.3.9=pypi_0
openssl=3.5.5=hbb43b14_0
openxlab=0.1.3=pypi_0
ordered-set=4.1.0=pypi_0
oss2=2.17.0=pypi_0
packaging=24.2=pypi_0
pandas=2.0.3=pypi_0
pillow=9.5.0=pypi_0
pip=24.2=py38haa95532_0
platformdirs=4.3.6=pypi_0
pycocotools=2.0.7=pypi_0
pycparser=2.23=pypi_0
pycryptodome=3.23.0=pypi_0
pygments=2.19.2=pypi_0
pyparsing=3.1.4=pypi_0
python=3.8.20=h8205438_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=1.8.0=py3.8_cuda10.2_cudnn7_0
pytz=2023.4=pypi_0
pywin32=311=pypi_0
pyyaml=6.0.3=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rich=13.4.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=60.2.0=pypi_0
six=1.17.0=pypi_0
sqlite=3.51.2=hee5a0db_0
sympy=1.13.3=pypi_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=h59b6b97_0
termcolor=2.4.0=pypi_0
terminaltables=3.1.10=pypi_0
tk=8.6.15=hf199647_0
tomli=2.4.0=pypi_0
torchvision=0.9.0=py38_cu102
tqdm=4.65.2=pypi_0
typing-extensions=4.13.2=pypi_0
tzdata=2025.3=pypi_0
ucrt=10.0.22621.0=haa95532_0
urllib3=1.26.20=pypi_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
wheel=0.44.0=py38haa95532_0
xz=5.8.2=h53af0af_0
yapf=0.43.0=pypi_0
zipp=3.20.2=pypi_0
zlib=1.3.1=h02ab6af_0
zstd=1.5.7=h56299aa_0
